Projects/LDTS/src/main/java/com/project/l02gr08/model/Position.java
package com.project.l02gr08.model;

import java.util.Objects;

public class Position {
    private final int x;
    private final int y;

    public Position(int xs, int ys) {
        x = xs;
        y = ys;
    }

    public Position(Position position) {
        x = position.getX();
        y = position.getY();
    }

    public int getX() {
        return x;
    }

    public int getY() {
        return y;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o)
            return true;
        if (!(o instanceof Position that))
            return false;
        return x == that.x && y == that.y;
    }

    @Override
    public int hashCode() {
        return Objects.hash(x, y);
    }

    public Position getLeft() {
        return new Position(x - 1, y);
    }

    public Position getRight() {
        return new Position(x + 1, y);
    }

    public Position getUp() {
        return new Position(x, y - 1);
    }

    public Position getDown() {
        return new Position(x, y + 1);
    }
}
